feat: restore canonical and alternate topics
Alternate topics let a single update carry a shared canonical topic
plus per-subscriber alternates, so one publish can serve several
private audiences at once instead of one publish per audience. The
OAuth-authorization rewrite dropped this in favor of a single topic
per update, but the removal traded away real capability for no actual
gain: the hub-enforced grant check and the transport storage model
both already handled multiple topics per update fine.

Re-enable it as a permanent, non-deprecated feature evaluated through
the current authorization_details/matcher-type model: a publisher
must be granted on every topic of an update, a subscriber needs a
grant on just one. Update the spec and docs to match.
