+++ publish.go
@@ -211,15 +211,6 @@ func (h *Hub) PublishHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// The protocol allows exactly one topic per update; alternate topics are
-	// a v8 feature, available only under the deprecated_topic build tag and
-	// WithProtocolVersionCompatibility(8).
-	if len(topics) > 1 && !h.allowsAlternateTopics() {
-		http.Error(w, `Multiple "topic" parameters are not supported anymore, publish one update per topic`, http.StatusBadRequest)
-
-		return
-	}
-
 	// Reject oversized topic lists before running canDispatch — otherwise
 	// an authenticated publisher could force O(topics × matchers)
 	// matching work on every request before being rejected by validate.
